requirements.txt
dj-database-url==0.5.0
Django==2.2.6
psycopg2==2.8.3
pytz==2019.3
djangorestframework==3.10.3
django-braces==1.13.0
django-rest-framework==0.1.0
django-cors-headers==3.1.1
django-redis==4.10.0
gunicorn==20.0.4
PyJWT==1.7.1
redis==3.4.1
sqlparse==0.3.1
whitenoise==5.0.1
